Moved each setting into its own file. Removed duplicate code for x and y axis
