+++ src/_constants.py
@@ -3,3 +3,4 @@ HIERARCHY_FILENAME = "ukbb_data_field_hierarchy.csv"
 STORAGE_BUCKET = "biobank-visualisation.appspot.com"
 
 SEX_FIELD_ID = '31'
+MAX_SELECTIONS = 30
